# Always permute on the device