--- a/src/details/ArborX_DetailsSortUtils.hpp
+++ b/src/details/ArborX_DetailsSortUtils.hpp
@@ -162,7 +162,7 @@ struct CopyOp<DstViewType, SrcViewType, 3>
 
 template <typename ExecutionSpace, typename PermutationView, typename InputView,
           typename OutputView>
-void applyReversePermutation(ExecutionSpace const &space,
+void applyInversePermutation(ExecutionSpace const &space,
                              PermutationView const &permutation,
                              InputView const &input_view,
                              OutputView const &output_view)
@@ -173,7 +173,7 @@ void applyReversePermutation(ExecutionSpace const &space,
   ARBORX_ASSERT(output_view.extent(0) == input_view.extent(0));
 
   Kokkos::parallel_for(
-      ARBORX_MARK_REGION("permute"),
+      ARBORX_MARK_REGION("inverse_permute"),
       Kokkos::RangePolicy<ExecutionSpace>(space, 0, input_view.extent(0)),
       KOKKOS_LAMBDA(int i) {
         PermuteHelper::CopyOp<OutputView, InputView>::copy(