@aprokop 2020-02-18
What are the number of values and queries for the posted results? What are the implications for very few of each (like 10^3 or 10^4)?