@masterleinad 2020-02-18
It's 10^6 points, 10^5 queries, 20 neighbors.