@masterleinad 2020-02-18
For 768 MPI processes, 20 neighbors, equal number of points and queries
Radius:

|  # points | old     |  new    |
|-----------|---------|---------|
|   1e2     | 1.08e-1 | 1.32e-1 |
|   1e3     | 1.35e-1 | 1.77e-1 |
|   1e4     | 2.50e-1 | 2.69e-1 |
|   1e5     | 5.24e-1 | 6.11e-1 |
|   1e6     | 2.99e0  | 2.42e0  |
|   1e7     | 2.66e1  | 2.04e1  |

knn:

|  # points | old     |  new    |
|-----------|---------|---------|
|   1e2     | 4.60e-1 | 5.75e-1 |
|   1e3     | 6.09e-1 | 8.52e-1 |
|   1e4     | 6.48e-1 | 9.70e-1 |
|   1e5     | 1.52e0  | 1.66e0  |
|   1e6     | 7.81e0  | 6.88e0  |
|   1e7     | 6.91e1  | 5.73e1  |

That is really confusing...