@aprokop 2020-02-18
So there is clearly a switch point around `1e5` where it becomes faster to permute on the device.