@aprokop 2020-02-18
> The only explanation I have is that launching the kernel takes more time than permuting in host space in serial.

Indeed.