@masterleinad 2020-02-18
But we only do a couple of kernel launches here and I can not imagine that this would make a difference in the order of 0.1s.