@masterleinad 2020-04-09
OK. I am getting pretty consistent results with what I had before running this time just with one process.

|  # points | old     |  new    |
|-----------|---------|---------|
|   1e2     | 3.754520e-02 | 4.117985e-02 |
|   1e3     | 4.997293e-02 | 5.488626e-02 |
|   1e4     | 1.083933e-01 | 1.068763e-01 |
|   1e5     | 6.154636e-01 | 5.275279e-01 |

knn:

|  # points | old     |  new    |
|-----------|---------|---------|
|   1e2     | 1.016623e-01 | 1.092507e-01
|   1e3     | 1.273524e-01 | 1.588901e-01
|   1e4     | 2.444361e-01 | 2.491620e-01
|   1e5     | 1.338598e+00 | 1.246120e+00

so around 1e4 it starts to get faster to do the sorting on the GPU.

For a large 1e7 queries and values the profiling looks like [old_profile.txt](https://github.com/arborx/ArborX/files/4459088/old_profile.txt) and [new_profile.txt](https://github.com/arborx/ArborX/files/4459089/new_profile.txt)
In particular, the total run time is 17.70298s for the new run and 26.68339 for the old one.

If we expect the normally to process more than 1e4--1e5 queries per process, we should go with the new one, otherwise sorting on the host is faster.