@masterleinad 2020-04-09
Ah, with profiling run times were only significantly faster with the current implementation for 1e2 queries.