@masterleinad 2020-04-14
Just checking the runtime for
```
  auto host_permute = time_monitor.getNewTimer("host_permute");
  host_permute->start();
  Kokkos::parallel_for(ARBORX_MARK_REGION("copy_destinations_permuted_host"),
                           Kokkos::RangePolicy<Kokkos::DefaultHostExecutionSpace>(
                               Kokkos::DefaultHostExecutionSpace{}, 0, n_values),
                           KOKKOS_LAMBDA(int const k) {
                             const int num_packets = 1;
                             int const i = k / num_packets;
                             int const j = k % num_packets;
                             dest_buffer_host(num_packets * permutation_vector_host[i] + j) =
                                 random_numbers_host[num_packets * i + j];
                           });
  Kokkos::fence();
  host_permute->stop();
  auto device_permute = time_monitor.getNewTimer("device_permute");
  device_permute->start();
      Kokkos::parallel_for(ARBORX_MARK_REGION("copy_destinations_permuted_device"),
                           Kokkos::RangePolicy<typename DeviceType::execution_space>(
                               typename DeviceType::execution_space{}, 0, n_values),
                           KOKKOS_LAMBDA(int const k) {
                             const int num_packets = 1;
                             int const i = k / num_packets;
                             int const j = k % num_packets;
                             dest_buffer_device(num_packets * permutation_vector_device[i] + j) =
                                 random_numbers_device[num_packets * i + j];
                           });
  Kokkos::fence();
  device_permute->stop();
```
gives me for (100 iterations)
```
|  # points | host     |  CUDA    |
|-----------|---------|---------|
|   1e2     | 3.593910e-04 | 1.967541e-03
|   1e3     | 5.597300e-04 | 1.850034e-03
|   1e4     | 5.199886e-03 | 1.976116e-03
|   1e5     | 1.399693e-01 | 2.783277e-03
|   1e6     | 2.096016e+00 | 2.081232e-02
```