@aprokop 2020-04-21
It seems that it is a common approach to offload permutations for small sizes to CPU. For example, PyTorch does

https://github.com/pytorch/pytorch/blob/4404762d7dd955383acee92e6f06b48144a0742e/aten/src/ATen/native/cuda/TensorFactories.cu#L85-L89

The corresponding issue pytorch/pytorch is 7606 (don't want to link).

I think it would be reasonable for us to do something similar. I'd like to see permute function isolated in Utils, which as switch to offload to CPU for small sizes. I think we could also use it for permutations in other parts. We have `applyPermutation` I think, that could also be updated.