@aprokop 2020-04-29
OK. The current version looks good. Just change reverse to inverse (or do that in a separate PR, as I think there are some other places), create an issue about potentially unifying implementation here with the one in batched queries (though, I'm not sure how), and rerun experiments (if it's not too much effort, I'm OK with merging without that).